feat(ml model): updating view of ml model feature list (#6576)
* updating view of ml model feature list

* Update TableOfMlFeatures.tsx
