+++ datahub-web-react/src/app/entity/mlModel/profile/MlModelFeaturesTab.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
 
-import { EntityType } from '../../../../types.generated';
+import { MlPrimaryKey, MlFeature } from '../../../../types.generated';
 import { useBaseEntity } from '../../shared/EntityContext';
 import { GetMlModelQuery } from '../../../../graphql/mlModel.generated';
-import { EntityList } from '../../shared/tabs/Entity/components/EntityList';
+import TableOfMlFeatures from '../../mlFeatureTable/profile/features/TableOfMlFeatures';
 
 export default function MlModelFeaturesTab() {
     const entity = useBaseEntity() as GetMlModelQuery;
 
     const model = entity && entity.mlModel;
-    const features = model?.features?.relationships.map((relationship) => relationship.entity);
+    const features = model?.features?.relationships.map((relationship) => relationship.entity) as Array<
+        MlFeature | MlPrimaryKey
+    >;
 
-    return <EntityList title="Features used" type={EntityType.Mlfeature} entities={features || []} />;
+    return <TableOfMlFeatures features={features || []} />;
 }
